refactor: extract zero-agents ops into zero-compose-service
Move updateComposeMetadata() and listComposes() from compose-service.ts
into a new zero-compose-service.ts, eliminating zero_agents table coupling
from the infra layer.

Closes #7678
